@@ -10,11 +10,11 @@
     </parent>
     <groupId>com.codeages</groupId>
     <artifactId>termius-plus</artifactId>
-    <version>v1.8.0</version>
+    <version>v1.8.2</version>
     <name>Termius Plus</name>
     <description>Termius Plus</description>
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <org.mapstruct.version>1.4.2.Final</org.mapstruct.version>
         <org.projectlombok.version>1.18.34</org.projectlombok.version>
         <org.apache.maven.compiler-plugin>3.13.0</org.apache.maven.compiler-plugin>
