@@ -3,6 +3,7 @@ package com.codeages.termiusplus.common.cache;
 import cn.hutool.core.convert.Convert;
 import cn.hutool.core.lang.TypeReference;
 import org.springframework.core.env.Environment;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
 import java.time.Duration;
@@ -13,14 +14,14 @@ import java.util.stream.Collectors;
 @Component
 public class RedisCacheManager {
 
-//    private final RedisTemplate<String, Object> redis;
+    private final RedisTemplate<String, Object> redis;
 
     private final String globalNamespace;
 
     private final Duration globalTTl;
 
-    public RedisCacheManager(Environment environment) {
-//        this.redis = redis;
+    public RedisCacheManager(Environment environment, RedisTemplate<String, Object> redis) {
+        this.redis = redis;
 
         globalNamespace = environment.getProperty("cache.namespace", "cache");
         var ttl = Convert.toLong(environment.getProperty("cache.ttl"), 36000L);
@@ -32,24 +33,23 @@ public class RedisCacheManager {
     }
 
     public <T> Optional<T> get(String namespace, String key, CacheGetCallable<T> callable, Duration ttl) {
-//        var obj = redis.opsForValue().get(makeKey(namespace, key));
-//        if ("NULL".equals(obj)) {
-//            return Optional.empty();
-//        }
-//
-//        if (obj != null) {
-//            var convertedObj = Convert.convert(new TypeReference<T>() {}, obj);
-//            if (convertedObj != null) {
-//                return Optional.of(convertedObj);
-//            }
-//        }
-//
-//        var objOp = callable.get();
-//
-//        redis.opsForValue().set(makeKey(namespace, key), objOp.isEmpty() ? "NULL" : objOp.get(), ttl);
-//
-//        return objOp;
-        return  null;
+        var obj = redis.opsForValue().get(makeKey(namespace, key));
+        if ("NULL".equals(obj)) {
+            return Optional.empty();
+        }
+
+        if (obj != null) {
+            var convertedObj = Convert.convert(new TypeReference<T>() {}, obj);
+            if (convertedObj != null) {
+                return Optional.of(convertedObj);
+            }
+        }
+
+        var objOp = callable.get();
+
+        redis.opsForValue().set(makeKey(namespace, key), objOp.isEmpty() ? "NULL" : objOp.get(), ttl);
+
+        return objOp;
     }
 
     public <T> Optional<T> getById(String namespace, Long id, CacheGetCallable<T> callable) {
@@ -65,44 +65,43 @@ public class RedisCacheManager {
     }
 
     public <T extends IdObject> Optional<T> getByRel(String namespace, String relKey, CacheGetCallable<T> callable, Duration ttl) {
-//        var idObj = redis.opsForValue().get(makeKey(namespace, relKey));
-//        if ("NULL".equals(idObj)) {
-//            return Optional.empty();
-//        }
-//
-//        var id = Convert.toLong(idObj);
-//
-//        Optional<T> objOp;
-//        var fromCache = false;
-//        if (id == null) {
-//            objOp = callable.get();
-//        } else {
-//            var obj = redis.opsForValue().get(makeKey(namespace, "id/"+id));
-//            if (obj == null || "NULL".equals(obj)) {
-//                objOp = callable.get();
-//            } else {
-//                var convertedObj = Convert.convert(new TypeReference<T>() {}, obj);
-//                if (convertedObj == null) {
-//                    objOp = callable.get();
-//                } else {
-//                    objOp = Optional.of(convertedObj);
-//                    fromCache = true;
-//                }
-//            }
-//        }
-//
-//        if (!fromCache) {
-//            if (objOp.isPresent()) {
-//                var obj = objOp.get();
-//                redis.opsForValue().set(makeKey(namespace, "id/"+obj.getId()), obj, ttl);
-//                redis.opsForValue().set(makeKey(namespace, relKey), obj.getId(), ttl);
-//            } else {
-//                redis.opsForValue().set(makeKey(namespace, relKey), "NULL", ttl);
-//            }
-//        }
-//
-//        return objOp;
-        return null;
+        var idObj = redis.opsForValue().get(makeKey(namespace, relKey));
+        if ("NULL".equals(idObj)) {
+            return Optional.empty();
+        }
+
+        var id = Convert.toLong(idObj);
+
+        Optional<T> objOp;
+        var fromCache = false;
+        if (id == null) {
+            objOp = callable.get();
+        } else {
+            var obj = redis.opsForValue().get(makeKey(namespace, "id/"+id));
+            if (obj == null || "NULL".equals(obj)) {
+                objOp = callable.get();
+            } else {
+                var convertedObj = Convert.convert(new TypeReference<T>() {}, obj);
+                if (convertedObj == null) {
+                    objOp = callable.get();
+                } else {
+                    objOp = Optional.of(convertedObj);
+                    fromCache = true;
+                }
+            }
+        }
+
+        if (!fromCache) {
+            if (objOp.isPresent()) {
+                var obj = objOp.get();
+                redis.opsForValue().set(makeKey(namespace, "id/"+obj.getId()), obj, ttl);
+                redis.opsForValue().set(makeKey(namespace, relKey), obj.getId(), ttl);
+            } else {
+                redis.opsForValue().set(makeKey(namespace, relKey), "NULL", ttl);
+            }
+        }
+
+        return objOp;
     }
 
     public void set(String namespace, String key, Object val) {
@@ -110,16 +109,16 @@ public class RedisCacheManager {
     }
 
     public void set(String namespace, String key, Object val, Duration ttl) {
-//        redis.opsForValue().set(makeKey(namespace, key), val, ttl);
+        redis.opsForValue().set(makeKey(namespace, key), val, ttl);
     }
 
     public void delete(String namespace, String key) {
-//        redis.delete(makeKey(namespace, key));
+        redis.delete(makeKey(namespace, key));
     }
 
     public void delete(String namespace, List<String> keys) {
         var fullKeys = keys.stream().map(key -> globalNamespace + "::" + namespace + "::" + key).collect(Collectors.toList());
-//        redis.delete(fullKeys);
+        redis.delete(fullKeys);
     }
 
     private String makeKey(String namespace, String key) {
